WIP: Add recent items and clipboard history to header bar menu (20)
